requirements.txt
tkinter as tk
numpy as np
os
shutil
matplotlib.pyplot as plt